Apply base classes to Pro version
