+++ source/CoordinateConversion/CoordinateConversionLibrary/ViewModels/CoordinateConversionViewModel.cs
@@ -12,14 +12,10 @@
   *   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. 
   *   See the License for the specific language governing permissions and 
   *   limitations under the License. 
-  ******************************************************************************/ 
+  ******************************************************************************/
 
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using CoordinateConversionLibrary.Models;
-using CoordinateConversionLibrary.Views;
-using CoordinateConversionLibrary.Helpers;
+
+//TODO delete this class/file
 
 namespace CoordinateConversionLibrary.ViewModels
 {
