+++ source/CoordinateConversion/ProAppCoordConversionModule/CoordinateConversionDockpaneViewModel.cs
@@ -32,42 +32,40 @@ using System.Collections.ObjectModel;
 using ArcGIS.Desktop.Mapping.Events;
 using ArcGIS.Core.Data;
 using System.Text.RegularExpressions;
+using ProAppCoordConversionModule.ViewModels;
 
 namespace ProAppCoordConversionModule
 {
     internal class CoordinateConversionDockpaneViewModel : DockPane
     {
-        private const string _dockPaneID = "ProAppCoordConversionModule_CoordinateConversionDockpane";
-
         protected CoordinateConversionDockpaneViewModel() 
         {
-            _coordinateConversionView = new CCConvertTabView();
-            HasInputError = false;
-            IsHistoryUpdate = true;
-            IsToolGenerated = false;
-            AddNewOCCommand = new CoordinateConversionLibrary.Helpers.RelayCommand(OnAddNewOCCommand);
-            ActivatePointToolCommand = new CoordinateConversionLibrary.Helpers.RelayCommand(OnMapToolCommand);
-            FlashPointCommand = new CoordinateConversionLibrary.Helpers.RelayCommand(OnFlashPointCommand);
-            CopyAllCommand = new CoordinateConversionLibrary.Helpers.RelayCommand(OnCopyAllCommand);
-            EditPropertiesDialogCommand = new CoordinateConversionLibrary.Helpers.RelayCommand(OnEditPropertiesDialogCommand);
-            Mediator.Register(CoordinateConversionLibrary.Constants.RequestCoordinateBroadcast, OnBCNeeded);
-            InputCoordinateHistoryList = new ObservableCollection<string>();
+            ConvertTabView = new CCConvertTabView();
+            ConvertTabView.DataContext = new ProConvertTabViewModel();
+
+            CollectTabView = new CCCollectTabView();
+            CollectTabView.DataContext = new ProCollectTabViewModel();
+
+            //_coordinateConversionView = new CCConvertTabView();
+            //HasInputError = false;
+            //IsHistoryUpdate = true;
+            //IsToolGenerated = false;
+            //ActivatePointToolCommand = new CoordinateConversionLibrary.Helpers.RelayCommand(OnMapToolCommand);
+            //FlashPointCommand = new CoordinateConversionLibrary.Helpers.RelayCommand(OnFlashPointCommand);
+            //EditPropertiesDialogCommand = new CoordinateConversionLibrary.Helpers.RelayCommand(OnEditPropertiesDialogCommand);
+            //Mediator.Register(CoordinateConversionLibrary.Constants.RequestCoordinateBroadcast, OnBCNeeded);
+            //InputCoordinateHistoryList = new ObservableCollection<string>();
             MapSelectionChangedEvent.Subscribe(OnSelectionChanged);
 
-            //var ctvm = CTView.Resources["CTViewModel"] as CoordinateConversionViewModel;
-            //if (ctvm != null)
-            //{
-            //    ctvm.SetCoordinateGetter(proCoordGetter);
-            //}
-            Mediator.NotifyColleagues(CoordinateConversionLibrary.Constants.SetCoordinateGetter, proCoordGetter);
+            //Mediator.NotifyColleagues(CoordinateConversionLibrary.Constants.SetCoordinateGetter, proCoordGetter);
 
-            configObserver = new PropertyObserver<CoordinateConversionLibraryConfig>(CoordinateConversionLibraryConfig.AddInConfig)
-            .RegisterHandler(n => n.DisplayCoordinateType, n => {
-                if(proCoordGetter != null && proCoordGetter.Point != null)
-                {
-                    InputCoordinate = proCoordGetter.GetInputDisplayString();
-                }
-            });
+            //configObserver = new PropertyObserver<CoordinateConversionLibraryConfig>(CoordinateConversionLibraryConfig.AddInConfig)
+            //.RegisterHandler(n => n.DisplayCoordinateType, n => {
+            //    if(proCoordGetter != null && proCoordGetter.Point != null)
+            //    {
+            //        InputCoordinate = proCoordGetter.GetInputDisplayString();
+            //    }
+            //});
         }
 
         ~CoordinateConversionDockpaneViewModel()
@@ -75,461 +73,455 @@ namespace ProAppCoordConversionModule
             MapSelectionChangedEvent.Unsubscribe(OnSelectionChanged);
         }
 
-        PropertyObserver<CoordinateConversionLibraryConfig> configObserver;
+        private const string _dockPaneID = "ProAppCoordConversionModule_CoordinateConversionDockpane";
+        
+        public CCConvertTabView ConvertTabView { get; set; }
+        public CCCollectTabView CollectTabView { get; set; }
 
-        public bool IsToolActive
-        {
-            get
-            {
-                if (FrameworkApplication.CurrentTool != null)
-                    return FrameworkApplication.CurrentTool == "ProAppCoordConversionModule_CoordinateMapTool";
+        //PropertyObserver<CoordinateConversionLibraryConfig> configObserver;
 
-                return false;
-            }
-            set
-            {
-                if (value)
-                    OnMapToolCommand(null);
-                else
-                    FrameworkApplication.SetCurrentToolAsync(string.Empty);
+        //public bool IsToolActive
+        //{
+        //    get
+        //    {
+        //        if (FrameworkApplication.CurrentTool != null)
+        //            return FrameworkApplication.CurrentTool == "ProAppCoordConversionModule_CoordinateMapTool";
 
-                NotifyPropertyChanged(new PropertyChangedEventArgs("IsToolActive"));
-            }
-        }
+        //        return false;
+        //    }
+        //    set
+        //    {
+        //        if (value)
+        //            OnMapToolCommand(null);
+        //        else
+        //            FrameworkApplication.SetCurrentToolAsync(string.Empty);
 
-        public ObservableCollection<string> InputCoordinateHistoryList { get; set; }
+        //        NotifyPropertyChanged(new PropertyChangedEventArgs("IsToolActive"));
+        //    }
+        //}
 
-        private void BroadcastCoordinateValues(MapPoint mapPoint)
-        {
-            var dict = new Dictionary<CoordinateType, string>();
-            if (mapPoint == null)
-                return;
+        //public ObservableCollection<string> InputCoordinateHistoryList { get; set; }
 
-            var dd = new CoordinateDD(mapPoint.Y, mapPoint.X);
+        //private void BroadcastCoordinateValues(MapPoint mapPoint)
+        //{
+        //    var dict = new Dictionary<CoordinateType, string>();
+        //    if (mapPoint == null)
+        //        return;
 
-            try
-            {
-                dict.Add(CoordinateType.DD, dd.ToString("", new CoordinateDDFormatter()));
-            }
-            catch { }
-            try
-            {
-                dict.Add(CoordinateType.DDM, new CoordinateDDM(dd).ToString("", new CoordinateDDMFormatter()));
-            }
-            catch { }
-            try
-            {
-                dict.Add(CoordinateType.DMS, new CoordinateDMS(dd).ToString("", new CoordinateDMSFormatter()));
-            }
-            catch { }
+        //    var dd = new CoordinateDD(mapPoint.Y, mapPoint.X);
 
-            Mediator.NotifyColleagues(CoordinateConversionLibrary.Constants.BroadcastCoordinateValues, dict);
+        //    try
+        //    {
+        //        dict.Add(CoordinateType.DD, dd.ToString("", new CoordinateDDFormatter()));
+        //    }
+        //    catch { }
+        //    try
+        //    {
+        //        dict.Add(CoordinateType.DDM, new CoordinateDDM(dd).ToString("", new CoordinateDDMFormatter()));
+        //    }
+        //    catch { }
+        //    try
+        //    {
+        //        dict.Add(CoordinateType.DMS, new CoordinateDMS(dd).ToString("", new CoordinateDMSFormatter()));
+        //    }
+        //    catch { }
 
-        }
-        private static System.IDisposable _overlayObject = null;
+        //    Mediator.NotifyColleagues(CoordinateConversionLibrary.Constants.BroadcastCoordinateValues, dict);
 
-        private void ClearOverlay()
-        {
-            if (_overlayObject != null)
-            {
-                _overlayObject.Dispose();
-                _overlayObject = null;
-            }
-        }
+        //}
+        //private static System.IDisposable _overlayObject = null;
 
-        private ProCoordinateGet proCoordGetter = new ProCoordinateGet();
+        //private void ClearOverlay()
+        //{
+        //    if (_overlayObject != null)
+        //    {
+        //        _overlayObject.Dispose();
+        //        _overlayObject = null;
+        //    }
+        //}
 
-        private bool _hasInputError = false;
-        public bool HasInputError
-        {
-            get { return _hasInputError; }
-            set
-            {
-                _hasInputError = value;
-                NotifyPropertyChanged(new PropertyChangedEventArgs("HasInputError"));
-            }
-        }
+        //private ProCoordinateGet proCoordGetter = new ProCoordinateGet();
 
-        public CoordinateConversionLibrary.Helpers.RelayCommand AddNewOCCommand { get; set; }
-        public CoordinateConversionLibrary.Helpers.RelayCommand ActivatePointToolCommand { get; set; }
-        public CoordinateConversionLibrary.Helpers.RelayCommand FlashPointCommand { get; set; }
-        public CoordinateConversionLibrary.Helpers.RelayCommand CopyAllCommand { get; set; }
-        public CoordinateConversionLibrary.Helpers.RelayCommand EditPropertiesDialogCommand { get; set; }
+        //private bool _hasInputError = false;
+        //public bool HasInputError
+        //{
+        //    get { return _hasInputError; }
+        //    set
+        //    {
+        //        _hasInputError = value;
+        //        NotifyPropertyChanged(new PropertyChangedEventArgs("HasInputError"));
+        //    }
+        //}
 
-        public bool IsHistoryUpdate { get; set; }
-        public bool IsToolGenerated { get; set; }
+        //public CoordinateConversionLibrary.Helpers.RelayCommand AddNewOCCommand { get; set; }
+        //public CoordinateConversionLibrary.Helpers.RelayCommand ActivatePointToolCommand { get; set; }
+        //public CoordinateConversionLibrary.Helpers.RelayCommand FlashPointCommand { get; set; }
+        //public CoordinateConversionLibrary.Helpers.RelayCommand CopyAllCommand { get; set; }
+        //public CoordinateConversionLibrary.Helpers.RelayCommand EditPropertiesDialogCommand { get; set; }
 
-        private string _inputCoordinate;
-        public string InputCoordinate
-        {
-            get
-            {
-                return _inputCoordinate;
-            }
+        //public bool IsHistoryUpdate { get; set; }
+        //public bool IsToolGenerated { get; set; }
 
-            set
-            {
-                ClearOverlay();
+        //private string _inputCoordinate;
+        //public string InputCoordinate
+        //{
+        //    get
+        //    {
+        //        return _inputCoordinate;
+        //    }
 
-                if (string.IsNullOrWhiteSpace(value))
-                    return;
+        //    set
+        //    {
+        //        ClearOverlay();
 
-                _inputCoordinate = value;
-                var tempDD = ProcessInput(_inputCoordinate);
+        //        if (string.IsNullOrWhiteSpace(value))
+        //            return;
 
-                // update tool view model
-                //TODO update for refactor, use Mediator, etc
-                //var ctvm = CTView.Resources["CTViewModel"] as CoordinateConversionViewModel;
-                //if (ctvm != null)
-                //{
-                //    ctvm.SetCoordinateGetter(proCoordGetter);
-                //    ctvm.InputCoordinate = tempDD;
-                //    var formattedInputCoordinate = proCoordGetter.GetInputDisplayString();
-                //    // update history
-                //    if (IsHistoryUpdate)
-                //    {
-                //        if (IsToolGenerated)
-                //            UIHelpers.UpdateHistory(formattedInputCoordinate, InputCoordinateHistoryList);
-                //        else
-                //            UIHelpers.UpdateHistory(_inputCoordinate, InputCoordinateHistoryList);
-                //    }
-                //    // reset flags
-                //    IsHistoryUpdate = true;
-                //    IsToolGenerated = false;
+        //        _inputCoordinate = value;
+        //        var tempDD = ProcessInput(_inputCoordinate);
 
-                //    _inputCoordinate = formattedInputCoordinate;
-                //}
+        //        // update tool view model
+        //        //TODO update for refactor, use Mediator, etc
+        //        //var ctvm = CTView.Resources["CTViewModel"] as CoordinateConversionViewModel;
+        //        //if (ctvm != null)
+        //        //{
+        //        //    ctvm.SetCoordinateGetter(proCoordGetter);
+        //        //    ctvm.InputCoordinate = tempDD;
+        //        //    var formattedInputCoordinate = proCoordGetter.GetInputDisplayString();
+        //        //    // update history
+        //        //    if (IsHistoryUpdate)
+        //        //    {
+        //        //        if (IsToolGenerated)
+        //        //            UIHelpers.UpdateHistory(formattedInputCoordinate, InputCoordinateHistoryList);
+        //        //        else
+        //        //            UIHelpers.UpdateHistory(_inputCoordinate, InputCoordinateHistoryList);
+        //        //    }
+        //        //    // reset flags
+        //        //    IsHistoryUpdate = true;
+        //        //    IsToolGenerated = false;
 
-                NotifyPropertyChanged(new PropertyChangedEventArgs("InputCoordinate"));
-            }
-        }
+        //        //    _inputCoordinate = formattedInputCoordinate;
+        //        //}
 
-        private CCConvertTabView _coordinateConversionView;
-        public CCConvertTabView CTView
-        {
-            get
-            {
-                return _coordinateConversionView;
-            }
-            set
-            {
-                _coordinateConversionView = value;
-            }
-        }
+        //        NotifyPropertyChanged(new PropertyChangedEventArgs("InputCoordinate"));
+        //    }
+        //}
+
+        //private CCConvertTabView _coordinateConversionView;
+        //public CCConvertTabView CTView
+        //{
+        //    get
+        //    {
+        //        return _coordinateConversionView;
+        //    }
+        //    set
+        //    {
+        //        _coordinateConversionView = value;
+        //    }
+        //}
 
         #region command handlers
 
-        private void OnAddNewOCCommand(object obj)
-        {
-            // Get name from user
-            string name = CoordinateType.DD.ToString();
-            Mediator.NotifyColleagues(CoordinateConversionLibrary.Constants.AddNewOutputCoordinate, new OutputCoordinateModel() { Name = name, CType = CoordinateType.DD, Format = "Y0.0#N X0.0#E" });
-        }
 
-        private void OnMapToolCommand(object obj)
-        {
-            FrameworkApplication.SetCurrentToolAsync("ProAppCoordConversionModule_CoordinateMapTool");
-        }
+        //private void OnMapToolCommand(object obj)
+        //{
+        //    FrameworkApplication.SetCurrentToolAsync("ProAppCoordConversionModule_CoordinateMapTool");
+        //}
 
-        private async void OnFlashPointCommand(object obj)
-        {
-            MapPoint point = null;
-            var previous = IsToolActive;
-            if (!IsToolActive)
-            {
-                IsToolActive = true;
-            }
+        //private async void OnFlashPointCommand(object obj)
+        //{
+        //    MapPoint point = null;
+        //    var previous = IsToolActive;
+        //    if (!IsToolActive)
+        //    {
+        //        IsToolActive = true;
+        //    }
 
-            //TODO fix this
-            //CoordinateDD dd;
-            //var ctvm = CTView.Resources["CTViewModel"] as CoordinateConversionViewModel;
-            //if (ctvm != null)
-            //{
-            //    if (!CoordinateDD.TryParse(ctvm.InputCoordinate, out dd))
-            //    {
-            //        Regex regexMercator = new Regex(@"^(?<latitude>\-?\d+\.?\d*)[+,;:\s]*(?<longitude>\-?\d+\.?\d*)");
+        //    //TODO fix this
+        //    //CoordinateDD dd;
+        //    //var ctvm = CTView.Resources["CTViewModel"] as CoordinateConversionViewModel;
+        //    //if (ctvm != null)
+        //    //{
+        //    //    if (!CoordinateDD.TryParse(ctvm.InputCoordinate, out dd))
+        //    //    {
+        //    //        Regex regexMercator = new Regex(@"^(?<latitude>\-?\d+\.?\d*)[+,;:\s]*(?<longitude>\-?\d+\.?\d*)");
 
-            //        var matchMercator = regexMercator.Match(ctvm.InputCoordinate);
+        //    //        var matchMercator = regexMercator.Match(ctvm.InputCoordinate);
 
-            //        if (matchMercator.Success && matchMercator.Length == ctvm.InputCoordinate.Length)
-            //        {
-            //            try
-            //            {
-            //                var Lat = Double.Parse(matchMercator.Groups["latitude"].Value);
-            //                var Lon = Double.Parse(matchMercator.Groups["longitude"].Value);
-            //                point = QueuedTask.Run(() =>
-            //                {
-            //                    return MapPointBuilder.CreateMapPoint(Lon, Lat, SpatialReferences.WebMercator);
-            //                }).Result;
-            //            }
-            //            catch (Exception ex)
-            //            {
-            //                // do nothing
-            //            }
-            //        }
-            //        else
-            //        {
-            //            return;
-            //        }
-            //    }
-            //}
-            //else { return; }
+        //    //        if (matchMercator.Success && matchMercator.Length == ctvm.InputCoordinate.Length)
+        //    //        {
+        //    //            try
+        //    //            {
+        //    //                var Lat = Double.Parse(matchMercator.Groups["latitude"].Value);
+        //    //                var Lon = Double.Parse(matchMercator.Groups["longitude"].Value);
+        //    //                point = QueuedTask.Run(() =>
+        //    //                {
+        //    //                    return MapPointBuilder.CreateMapPoint(Lon, Lat, SpatialReferences.WebMercator);
+        //    //                }).Result;
+        //    //            }
+        //    //            catch (Exception ex)
+        //    //            {
+        //    //                // do nothing
+        //    //            }
+        //    //        }
+        //    //        else
+        //    //        {
+        //    //            return;
+        //    //        }
+        //    //    }
+        //    //}
+        //    //else { return; }
 
-            ArcGIS.Core.CIM.CIMPointSymbol symbol = null;
+        //    ArcGIS.Core.CIM.CIMPointSymbol symbol = null;
 
-            //TODO fix this also
-            //if (point == null)
-            //{
-            //    point = await QueuedTask.Run(() =>
-            //    {
-            //        ArcGIS.Core.Geometry.SpatialReference sptlRef = SpatialReferenceBuilder.CreateSpatialReference(4326);
-            //        return MapPointBuilder.CreateMapPoint(dd.Lon, dd.Lat, sptlRef);
-            //    });
-            //}
+        //    //TODO fix this also
+        //    //if (point == null)
+        //    //{
+        //    //    point = await QueuedTask.Run(() =>
+        //    //    {
+        //    //        ArcGIS.Core.Geometry.SpatialReference sptlRef = SpatialReferenceBuilder.CreateSpatialReference(4326);
+        //    //        return MapPointBuilder.CreateMapPoint(dd.Lon, dd.Lat, sptlRef);
+        //    //    });
+        //    //}
 
-            ////
+        //    ////
 
-            //await QueuedTask.Run(() =>
-            //{
-            //    // Construct point symbol
-            //    symbol = SymbolFactory.ConstructPointSymbol(ColorFactory.Red, 10.0, SimpleMarkerStyle.Star);
-            //});
+        //    //await QueuedTask.Run(() =>
+        //    //{
+        //    //    // Construct point symbol
+        //    //    symbol = SymbolFactory.ConstructPointSymbol(ColorFactory.Red, 10.0, SimpleMarkerStyle.Star);
+        //    //});
 
-            ////Get symbol reference from the symbol 
-            //CIMSymbolReference symbolReference = symbol.MakeSymbolReference();
+        //    ////Get symbol reference from the symbol 
+        //    //CIMSymbolReference symbolReference = symbol.MakeSymbolReference();
 
-            //QueuedTask.Run(() =>
-            //{
-            //    ClearOverlay();
-            //    _overlayObject = MapView.Active.AddOverlay(point, symbolReference);
-            //    //MapView.Active.ZoomToAsync(point, new TimeSpan(2500000), true);
-            //});
+        //    //QueuedTask.Run(() =>
+        //    //{
+        //    //    ClearOverlay();
+        //    //    _overlayObject = MapView.Active.AddOverlay(point, symbolReference);
+        //    //    //MapView.Active.ZoomToAsync(point, new TimeSpan(2500000), true);
+        //    //});
 
 
-            await QueuedTask.Run(() =>
-                {
-                    // is point within current map extent
-                    var projectedPoint = GeometryEngine.Project(point, MapView.Active.Extent.SpatialReference);
-                    if(!GeometryEngine.Contains(MapView.Active.Extent, projectedPoint))
-                    {
-                        MapView.Active.PanTo(point);
-                    }
-                    Mediator.NotifyColleagues("UPDATE_FLASH", point);
-                });
+        //    await QueuedTask.Run(() =>
+        //        {
+        //            // is point within current map extent
+        //            var projectedPoint = GeometryEngine.Project(point, MapView.Active.Extent.SpatialReference);
+        //            if(!GeometryEngine.Contains(MapView.Active.Extent, projectedPoint))
+        //            {
+        //                MapView.Active.PanTo(point);
+        //            }
+        //            Mediator.NotifyColleagues("UPDATE_FLASH", point);
+        //        });
 
-            //await QueuedTask.Run(() =>
-            //{
-            //    Task.Delay(500);
-            //    ClearOverlay();
-            //    //_overlayObject = MapView.Active.AddOverlay(point, symbolReference);
-            //    //MapView.Active.ZoomToAsync(point, new TimeSpan(2500000), true);
-            //});
-            //if (previous != IsToolActive)
-            //    IsToolActive = previous;
-            //await QueuedTask.Run(() =>
-            //{
-            //    Task.Delay(500);
-            //    ClearOverlay();
-            //    MapView.Active.LookAt(MapView.Active.Extent.Center);
-            //});
-        }
+        //    //await QueuedTask.Run(() =>
+        //    //{
+        //    //    Task.Delay(500);
+        //    //    ClearOverlay();
+        //    //    //_overlayObject = MapView.Active.AddOverlay(point, symbolReference);
+        //    //    //MapView.Active.ZoomToAsync(point, new TimeSpan(2500000), true);
+        //    //});
+        //    //if (previous != IsToolActive)
+        //    //    IsToolActive = previous;
+        //    //await QueuedTask.Run(() =>
+        //    //{
+        //    //    Task.Delay(500);
+        //    //    ClearOverlay();
+        //    //    MapView.Active.LookAt(MapView.Active.Extent.Center);
+        //    //});
+        //}
 
-        private void OnCopyAllCommand(object obj)
-        {
-            Mediator.NotifyColleagues(CoordinateConversionLibrary.Constants.CopyAllCoordinateOutputs, InputCoordinate);
-        }
+        //private void OnEditPropertiesDialogCommand(object obj)
+        //{
+        //    var dlg = new EditPropertiesView();
 
-        private void OnEditPropertiesDialogCommand(object obj)
-        {
-            var dlg = new EditPropertiesView();
+        //    dlg.ShowDialog();
+        //}
 
-            dlg.ShowDialog();
-        }
+        //private void OnBCNeeded(object obj)
+        //{
+        //    if (proCoordGetter == null || proCoordGetter.Point == null)
+        //        return;
 
-        private void OnBCNeeded(object obj)
-        {
-            if (proCoordGetter == null || proCoordGetter.Point == null)
-                return;
-
-            BroadcastCoordinateValues(proCoordGetter.Point);
-        }
+        //    BroadcastCoordinateValues(proCoordGetter.Point);
+        //}
 
         #endregion command handlers
 
-        internal string GetFormattedCoordinate(string coord, CoordinateType cType)
-        {
-            string format = "";
+        //internal string GetFormattedCoordinate(string coord, CoordinateType cType)
+        //{
+        //    string format = "";
 
-            var tt = CoordinateConversionLibraryConfig.AddInConfig.OutputCoordinateList.FirstOrDefault(t => t.CType == cType);
-            if (tt != null)
-            {
-                format = tt.Format;
-                Console.WriteLine(tt.Format);
-            }
+        //    var tt = CoordinateConversionLibraryConfig.AddInConfig.OutputCoordinateList.FirstOrDefault(t => t.CType == cType);
+        //    if (tt != null)
+        //    {
+        //        format = tt.Format;
+        //        Console.WriteLine(tt.Format);
+        //    }
 
-            var cf = CoordinateHandler.GetFormattedCoord(cType, coord, format);
+        //    var cf = CoordinateHandler.GetFormattedCoord(cType, coord, format);
 
-            if (!String.IsNullOrWhiteSpace(cf))
-                return cf;
+        //    if (!String.IsNullOrWhiteSpace(cf))
+        //        return cf;
 
-            return string.Empty;
-        }
+        //    return string.Empty;
+        //}
 
-        private string ProcessInput(string input)
-        {
-            string result = string.Empty;
-            MapPoint point;
-            HasInputError = false;
+        //private string ProcessInput(string input)
+        //{
+        //    string result = string.Empty;
+        //    MapPoint point;
+        //    HasInputError = false;
 
-            if (string.IsNullOrWhiteSpace(input))
-                return result;
+        //    if (string.IsNullOrWhiteSpace(input))
+        //        return result;
 
-            var coordType = GetCoordinateType(input, out point);
+        //    var coordType = GetCoordinateType(input, out point);
 
-            if (coordType == CoordinateType.Unknown)
-                HasInputError = true;
-            else
-            {
-                proCoordGetter.Point = point;
-                result = new CoordinateDD(point.Y, point.X).ToString("", new CoordinateDDFormatter());
-            }
+        //    if (coordType == CoordinateType.Unknown)
+        //        HasInputError = true;
+        //    else
+        //    {
+        //        proCoordGetter.Point = point;
+        //        result = new CoordinateDD(point.Y, point.X).ToString("", new CoordinateDDFormatter());
+        //    }
 
-            return result;
-        }
+        //    return result;
+        //}
 
-        private CoordinateType GetCoordinateType(string input, out MapPoint point)
-        {
-            point = null;
+        //private CoordinateType GetCoordinateType(string input, out MapPoint point)
+        //{
+        //    point = null;
 
-            // DD
-            CoordinateDD dd;
-            if(CoordinateDD.TryParse(input, out dd))
-            {
-                point = QueuedTask.Run(() =>
-                {
-                    ArcGIS.Core.Geometry.SpatialReference sptlRef = SpatialReferenceBuilder.CreateSpatialReference(4326);
-                    return MapPointBuilder.CreateMapPoint(dd.Lon, dd.Lat, sptlRef);
-                }).Result;
-                return CoordinateType.DD;
-            }
+        //    // DD
+        //    CoordinateDD dd;
+        //    if(CoordinateDD.TryParse(input, out dd))
+        //    {
+        //        point = QueuedTask.Run(() =>
+        //        {
+        //            ArcGIS.Core.Geometry.SpatialReference sptlRef = SpatialReferenceBuilder.CreateSpatialReference(4326);
+        //            return MapPointBuilder.CreateMapPoint(dd.Lon, dd.Lat, sptlRef);
+        //        }).Result;
+        //        return CoordinateType.DD;
+        //    }
 
-            // DDM
-            CoordinateDDM ddm;
-            if(CoordinateDDM.TryParse(input, out ddm))
-            {
-                dd = new CoordinateDD(ddm);
-                point = QueuedTask.Run(() =>
-                {
-                    ArcGIS.Core.Geometry.SpatialReference sptlRef = SpatialReferenceBuilder.CreateSpatialReference(4326);
-                    return MapPointBuilder.CreateMapPoint(dd.Lon, dd.Lat, sptlRef);
-                }).Result;
-                return CoordinateType.DDM;
-            }
-            // DMS
-            CoordinateDMS dms;
-            if (CoordinateDMS.TryParse(input, out dms))
-            {
-                dd = new CoordinateDD(dms);
-                point = QueuedTask.Run(() =>
-                {
-                    ArcGIS.Core.Geometry.SpatialReference sptlRef = SpatialReferenceBuilder.CreateSpatialReference(4326);
-                    return MapPointBuilder.CreateMapPoint(dd.Lon, dd.Lat, sptlRef);
-                }).Result;
-                return CoordinateType.DMS;
-            }
+        //    // DDM
+        //    CoordinateDDM ddm;
+        //    if(CoordinateDDM.TryParse(input, out ddm))
+        //    {
+        //        dd = new CoordinateDD(ddm);
+        //        point = QueuedTask.Run(() =>
+        //        {
+        //            ArcGIS.Core.Geometry.SpatialReference sptlRef = SpatialReferenceBuilder.CreateSpatialReference(4326);
+        //            return MapPointBuilder.CreateMapPoint(dd.Lon, dd.Lat, sptlRef);
+        //        }).Result;
+        //        return CoordinateType.DDM;
+        //    }
+        //    // DMS
+        //    CoordinateDMS dms;
+        //    if (CoordinateDMS.TryParse(input, out dms))
+        //    {
+        //        dd = new CoordinateDD(dms);
+        //        point = QueuedTask.Run(() =>
+        //        {
+        //            ArcGIS.Core.Geometry.SpatialReference sptlRef = SpatialReferenceBuilder.CreateSpatialReference(4326);
+        //            return MapPointBuilder.CreateMapPoint(dd.Lon, dd.Lat, sptlRef);
+        //        }).Result;
+        //        return CoordinateType.DMS;
+        //    }
 
-            CoordinateGARS gars;
-            if (CoordinateGARS.TryParse(input, out gars))
-            {
-                try
-                {
-                    point = QueuedTask.Run(() =>
-                    {
-                        ArcGIS.Core.Geometry.SpatialReference sptlRef = SpatialReferenceBuilder.CreateSpatialReference(4326);
-                        var tmp = MapPointBuilder.FromGeoCoordinateString(gars.ToString("", new CoordinateGARSFormatter()), sptlRef, GeoCoordinateType.GARS, FromGeoCoordinateMode.Default);
-                        return tmp;
-                    }).Result;
+        //    CoordinateGARS gars;
+        //    if (CoordinateGARS.TryParse(input, out gars))
+        //    {
+        //        try
+        //        {
+        //            point = QueuedTask.Run(() =>
+        //            {
+        //                ArcGIS.Core.Geometry.SpatialReference sptlRef = SpatialReferenceBuilder.CreateSpatialReference(4326);
+        //                var tmp = MapPointBuilder.FromGeoCoordinateString(gars.ToString("", new CoordinateGARSFormatter()), sptlRef, GeoCoordinateType.GARS, FromGeoCoordinateMode.Default);
+        //                return tmp;
+        //            }).Result;
                     
-                    return CoordinateType.GARS;
-                }
-                catch { }
-            }
+        //            return CoordinateType.GARS;
+        //        }
+        //        catch { }
+        //    }
 
-            CoordinateMGRS mgrs;
-            if(CoordinateMGRS.TryParse(input, out mgrs))
-            {
-                try
-                {
-                    point = QueuedTask.Run(() =>
-                    {
-                        ArcGIS.Core.Geometry.SpatialReference sptlRef = SpatialReferenceBuilder.CreateSpatialReference(4326);
-                        var tmp = MapPointBuilder.FromGeoCoordinateString(mgrs.ToString("", new CoordinateMGRSFormatter()), sptlRef, GeoCoordinateType.MGRS, FromGeoCoordinateMode.Default);
-                        return tmp;
-                    }).Result;
+        //    CoordinateMGRS mgrs;
+        //    if(CoordinateMGRS.TryParse(input, out mgrs))
+        //    {
+        //        try
+        //        {
+        //            point = QueuedTask.Run(() =>
+        //            {
+        //                ArcGIS.Core.Geometry.SpatialReference sptlRef = SpatialReferenceBuilder.CreateSpatialReference(4326);
+        //                var tmp = MapPointBuilder.FromGeoCoordinateString(mgrs.ToString("", new CoordinateMGRSFormatter()), sptlRef, GeoCoordinateType.MGRS, FromGeoCoordinateMode.Default);
+        //                return tmp;
+        //            }).Result;
 
-                    return CoordinateType.MGRS;
-                }
-                catch { }
-            }
+        //            return CoordinateType.MGRS;
+        //        }
+        //        catch { }
+        //    }
 
-            CoordinateUSNG usng;
-            if (CoordinateUSNG.TryParse(input, out usng))
-            {
-                try
-                {
-                    point = QueuedTask.Run(() =>
-                    {
-                        ArcGIS.Core.Geometry.SpatialReference sptlRef = SpatialReferenceBuilder.CreateSpatialReference(4326);
-                        var tmp = MapPointBuilder.FromGeoCoordinateString(usng.ToString("", new CoordinateMGRSFormatter()), sptlRef, GeoCoordinateType.USNG, FromGeoCoordinateMode.Default);
-                        return tmp;
-                    }).Result;
+        //    CoordinateUSNG usng;
+        //    if (CoordinateUSNG.TryParse(input, out usng))
+        //    {
+        //        try
+        //        {
+        //            point = QueuedTask.Run(() =>
+        //            {
+        //                ArcGIS.Core.Geometry.SpatialReference sptlRef = SpatialReferenceBuilder.CreateSpatialReference(4326);
+        //                var tmp = MapPointBuilder.FromGeoCoordinateString(usng.ToString("", new CoordinateMGRSFormatter()), sptlRef, GeoCoordinateType.USNG, FromGeoCoordinateMode.Default);
+        //                return tmp;
+        //            }).Result;
 
-                    return CoordinateType.USNG;
-                }
-                catch { }
-            }
+        //            return CoordinateType.USNG;
+        //        }
+        //        catch { }
+        //    }
 
-            CoordinateUTM utm;
-            if (CoordinateUTM.TryParse(input, out utm))
-            {
-                try
-                {
-                    point = QueuedTask.Run(() =>
-                    {
-                        ArcGIS.Core.Geometry.SpatialReference sptlRef = SpatialReferenceBuilder.CreateSpatialReference(4326);
-                        var tmp = MapPointBuilder.FromGeoCoordinateString(utm.ToString("", new CoordinateUTMFormatter()), sptlRef, GeoCoordinateType.UTM, FromGeoCoordinateMode.Default);
-                        return tmp;
-                    }).Result;
+        //    CoordinateUTM utm;
+        //    if (CoordinateUTM.TryParse(input, out utm))
+        //    {
+        //        try
+        //        {
+        //            point = QueuedTask.Run(() =>
+        //            {
+        //                ArcGIS.Core.Geometry.SpatialReference sptlRef = SpatialReferenceBuilder.CreateSpatialReference(4326);
+        //                var tmp = MapPointBuilder.FromGeoCoordinateString(utm.ToString("", new CoordinateUTMFormatter()), sptlRef, GeoCoordinateType.UTM, FromGeoCoordinateMode.Default);
+        //                return tmp;
+        //            }).Result;
 
-                    return CoordinateType.UTM;
-                }
-                catch { }
-            }
+        //            return CoordinateType.UTM;
+        //        }
+        //        catch { }
+        //    }
 
-            Regex regexMercator = new Regex(@"^(?<latitude>\-?\d+\.?\d*)[+,;:\s]*(?<longitude>\-?\d+\.?\d*)");
+        //    Regex regexMercator = new Regex(@"^(?<latitude>\-?\d+\.?\d*)[+,;:\s]*(?<longitude>\-?\d+\.?\d*)");
 
-            var matchMercator = regexMercator.Match(input);
+        //    var matchMercator = regexMercator.Match(input);
 
-            if (matchMercator.Success && matchMercator.Length == input.Length)
-            {
-                try
-                {
-                    var Lat = Double.Parse(matchMercator.Groups["latitude"].Value);
-                    var Lon = Double.Parse(matchMercator.Groups["longitude"].Value);
-                    point = QueuedTask.Run(() =>
-                    {
-                        return MapPointBuilder.CreateMapPoint(Lon, Lat, SpatialReferences.WebMercator);
-                    }).Result;
-                    return CoordinateType.DD;
-                }
-                catch (Exception ex)
-                {
-                    return CoordinateType.Unknown;
-                }
-            }
+        //    if (matchMercator.Success && matchMercator.Length == input.Length)
+        //    {
+        //        try
+        //        {
+        //            var Lat = Double.Parse(matchMercator.Groups["latitude"].Value);
+        //            var Lon = Double.Parse(matchMercator.Groups["longitude"].Value);
+        //            point = QueuedTask.Run(() =>
+        //            {
+        //                return MapPointBuilder.CreateMapPoint(Lon, Lat, SpatialReferences.WebMercator);
+        //            }).Result;
+        //            return CoordinateType.DD;
+        //        }
+        //        catch (Exception ex)
+        //        {
+        //            return CoordinateType.Unknown;
+        //        }
+        //    }
 
-            return CoordinateType.Unknown;
-        }
+        //    return CoordinateType.Unknown;
+        //}
 
         /// <summary>
         /// Show the DockPane.
@@ -558,7 +550,7 @@ namespace ProAppCoordConversionModule
                     {
                         var SelectedOID = fl.GetSelection().GetObjectIDs().FirstOrDefault();
                         if (SelectedOID < 0)
-                            return string.Empty;
+                            return null;
 
                         var SelectedLayer = fl as BasicFeatureLayer;
 
@@ -571,7 +563,7 @@ namespace ProAppCoordConversionModule
                             row = cursor.Current;
 
                         if (row == null)
-                            return string.Empty;
+                            return null;
 
                         var fields = row.GetFields();
                         lock (_lock)
@@ -585,10 +577,11 @@ namespace ProAppCoordConversionModule
                                     if (val is MapPoint)
                                     {
                                         var temp = val as MapPoint;
-                                        // project to WGS1984
-                                        proCoordGetter.Point = temp;
-                                        proCoordGetter.Project(4326);
-                                        return string.Format("{0:0.0####} {1:0.0####}", proCoordGetter.Point.Y, proCoordGetter.Point.X);
+                                        return temp;
+                                        //// project to WGS1984
+                                        //proCoordGetter.Point = temp;
+                                        //proCoordGetter.Project(4326);
+                                        //return string.Format("{0:0.0####} {1:0.0####}", proCoordGetter.Point.Y, proCoordGetter.Point.X);
                                     }
                                     break;
                                 }
@@ -599,13 +592,16 @@ namespace ProAppCoordConversionModule
                     {
                     }
 
-                    return string.Empty;
+                    return null;
                 });
 
-                if (!string.IsNullOrWhiteSpace(pointd))
-                {
-                    InputCoordinate = pointd;
-                }
+                if(pointd != null)
+                    Mediator.NotifyColleagues(CoordinateConversionLibrary.Constants.NewMapPointSelection, pointd);
+
+                //if (!string.IsNullOrWhiteSpace(pointd))
+                //{
+                //    InputCoordinate = pointd;
+                //}
             }
         }
