+++ source/CoordinateConversion/ProAppCoordConversionModule/ViewModels/ProCollectTabViewModel.cs
@@ -1,0 +1,17 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+//TODO implement this view model
+namespace ProAppCoordConversionModule.ViewModels
+{
+    public class ProCollectTabViewModel : ProTabBaseViewModel
+    {
+        public ProCollectTabViewModel()
+        {
+
+        }
+    }
+}
